Small internal refactor of ShaderProcessor
