+++ src/platform/graphics/uniform-buffer-format.js
@@ -1,7 +1,7 @@
 import { Debug } from '../../core/debug.js';
 import { math } from '../../core/math/math.js';
 import {
-    uniformTypeToName, bindGroupNames,
+    uniformTypeToName,
     UNIFORMTYPE_BOOL, UNIFORMTYPE_INT, UNIFORMTYPE_FLOAT, UNIFORMTYPE_UINT, UNIFORMTYPE_VEC2,
     UNIFORMTYPE_VEC3, UNIFORMTYPE_VEC4, UNIFORMTYPE_IVEC2, UNIFORMTYPE_UVEC2, UNIFORMTYPE_IVEC3,
     UNIFORMTYPE_IVEC4, UNIFORMTYPE_BVEC2, UNIFORMTYPE_BVEC3, UNIFORMTYPE_UVEC3, UNIFORMTYPE_BVEC4,
@@ -252,20 +252,6 @@ class UniformBufferFormat {
      */
     get(name) {
         return this.map.get(name);
-    }
-
-    getShaderDeclaration(bindGroup, bindIndex) {
-
-        const name = bindGroupNames[bindGroup];
-        let code = `layout(set = ${bindGroup}, binding = ${bindIndex}, std140) uniform ub_${name} {\n`;
-
-        this.uniforms.forEach((uniform) => {
-            const typeString = uniformTypeToName[uniform.type];
-            Debug.assert(typeString.length > 0, `Uniform type ${uniform.type} is not handled.`);
-            code += `    ${typeString} ${uniform.shortName}${uniform.count ? `[${uniform.count}]` : ''};\n`;
-        });
-
-        return `${code}};\n`;
     }
 }
