+++ src/platform/graphics/webgpu/webgpu-shader.js
@@ -1,7 +1,7 @@
 import { Debug, DebugHelper } from '../../../core/debug.js';
 import { SHADERLANGUAGE_WGSL } from '../constants.js';
 import { DebugGraphics } from '../debug-graphics.js';
-import { ShaderProcessor } from '../shader-processor.js';
+import { ShaderProcessorGLSL } from '../shader-processor-glsl.js';
 import { WebgpuDebug } from './webgpu-debug.js';
 import { WebgpuShaderProcessorWGSL } from './webgpu-shader-processor-wgsl.js';
 
@@ -143,7 +143,7 @@ class WebgpuShader {
         const shader = this.shader;
 
         // process the shader source to allow for uniforms
-        const processed = ShaderProcessor.run(shader.device, shader.definition, shader);
+        const processed = ShaderProcessorGLSL.run(shader.device, shader.definition, shader);
 
         // keep reference to processed shaders in debug mode
         Debug.call(() => {
